LU-5577 obdclass: cleanup comparison between signed and unsigned
Change arguments for page offset from int to loff_t.
Change type of counter and size variables to size_t.

+++ lustre/include/lustre/lustre_idl.h
@@ -1085,9 +1085,9 @@ static inline struct lu_dirent *lu_dirent_next(struct lu_dirent *ent)
         return next;
 }
 
-static inline int lu_dirent_calc_size(int namelen, __u16 attr)
+static inline size_t lu_dirent_calc_size(size_t namelen, __u16 attr)
 {
-        int size;
+	size_t size;
 
         if (attr & LUDA_TYPE) {
                 const unsigned align = sizeof(struct luda_type) - 1;
@@ -1099,7 +1099,7 @@ static inline int lu_dirent_calc_size(int namelen, __u16 attr)
         return (size + 7) & ~7;
 }
 
-static inline int lu_dirent_size(const struct lu_dirent *ent)
+static inline size_t lu_dirent_size(const struct lu_dirent *ent)
 {
         if (le16_to_cpu(ent->lde_reclen) == 0) {
                 return lu_dirent_calc_size(le16_to_cpu(ent->lde_namelen),
